+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Feature
 - Big token movers with subscription [#1191](https://github.com/EthVM/EthVM/pull/1191)
 
+###  Fix
+- fis loading state on portfolio list [#1191](https://github.com/EthVM/EthVM/pull/1191)
 
 ### UI
 - clean up color system [#1191](https://github.com/EthVM/EthVM/pull/1191)